Postgres/requirements.txt
fastapi
uvicorn
psycopg2-binary
playwright